refactor(chrome): 拆 1001 行 chrome.py + session.py 抽 transport 层
- cdp/commands/chrome.py（分类错误：实为进程启动器）拆为 cdp/browser_detection.py
  （BrowserType + 检测）+ cdp/launcher.py（ChromeLauncher）+ cdp/process.py
  （kill_existing_chrome 降为模块函数），移出 commands/，调用方 import 全更新。
- cdp/session.py 841→558：WS 传输/请求转发/事件多路复用抽到 transport.py 的
  CDPTransport 类（与既有函数式原语共存），session 退化为门面 + lazy property。
行为冻结，tests/extension parity 全绿。
