+++ src/frago/chrome/cdp/commands/input.py
@@ -4,7 +4,7 @@ Input-related CDP commands
 Encapsulates CDP commands for the Input domain.
 """
 
-from typing import Dict, Any, Optional
+from typing import Dict, Any
 
 from ..session import CDPSession
 from ..logger import get_logger
